+++ packages/encoding/src/index.ts
@@ -1,5 +1,11 @@
 // Auto-generated index file for encodingFunctions
 // This file exports all functions from this module
 
+export * from './decodeBase32';
 export * from './decodeBase64';
+export * from './decodeBase64URL';
+export * from './decodeHex';
+export * from './encodeBase32';
 export * from './encodeBase64';
+export * from './encodeBase64URL';
+export * from './encodeHex';